separate Mission tracker from flowchart
