+++ pipelex/mission/track/mission_tracker.py
@@ -19,62 +19,18 @@ from pipelex.config import get_config
 from pipelex.core.concept import Concept
 from pipelex.core.stuff import Stuff
 from pipelex.exceptions import JobHistoryError
-from pipelex.mission.mission_tracker_protocol import MissionTrackerProtocol
+from pipelex.mission.track.flow_chart import MissionFlowChart
+from pipelex.mission.track.mission_tracker_protocol import MissionTrackerProtocol
+from pipelex.mission.track.tracker_models import (
+    EdgeAttributeKey,
+    EdgeCategory,
+    NodeAttributeKey,
+    NodeCategory,
+    SpecialNodeName,
+    SubGraphClassDef,
+)
 from pipelex.pipe_controllers.pipe_condition_details import PipeConditionDetails
-from pipelex.tools.misc.mermaid_helpers import clean_str_for_mermaid_node_title, make_mermaid_url, print_mermaid_url
-from pipelex.tools.utils.string_utils import snake_to_capitalize_first_letter
-
-
-class NodeCategory(StrEnum):
-    SPECIAL = "special"
-    STUFF = "stuff"
-    CONDITION = "condition"
-
-
-class NodeAttributeKey(StrEnum):
-    CATEGORY = "category"
-    NAME = "name"
-    TAG = "tag"
-    DESCRIPTION = "description"
-    DEBUG_INFO = "debug_info"
-    SUBGRAPH = "subgraph"
-    COMMENT = "comment"
-
-
-class SpecialNodeName(StrEnum):
-    START = "start"
-
-
-class EdgeCategory(StrEnum):
-    PIPE = "pipe"
-    BATCH = "batch"
-    AGGREGATE = "aggregate"
-    CONDITION = "condition"
-    CHOICE = "choice"
-
-
-class EdgeAttributeKey(StrEnum):
-    EDGE_CATEGORY = "edge_category"
-    PIPE_CODE = "pipe_code"
-    CONDITION_EXPRESSION = "condition_expression"
-    CHOSEN_PIPE = "chosen_pipe"
-
-
-class GraphTree(BaseModel):
-    nodes_by_subgraph: Dict[str, List[str]]
-
-
-class SubGraphClassDef(BaseModel):
-    name: str
-    color: str
-
-
-def nice_edge_tag(edge_tag: str) -> str:
-    return f'"{snake_to_capitalize_first_letter(edge_tag)}"'
-
-
-def _indent_line(line: str, indent: int) -> str:
-    return f"{'    ' * indent}{line}"
+from pipelex.tools.misc.mermaid_helpers import print_mermaid_url
 
 
 class MissionTracker(MissionTrackerProtocol):
@@ -82,22 +38,14 @@ class MissionTracker(MissionTrackerProtocol):
         self.is_active: bool = False
         self.nx_graph: nx.DiGraph = nx.DiGraph()
         self.start_node: Optional[str] = None
-        self.sub_graph_class_defs: List[SubGraphClassDef] = []
 
     def activate(self):
         self.is_active = True
-
-        history_graph_config = get_config().pipelex.history_graph_config
-        for sub_graph_index, sub_graph_color in enumerate(history_graph_config.sub_graph_colors):
-            class_def_letter = chr(ord("a") + sub_graph_index)
-            class_def_name = f"sub_{class_def_letter}"
-            self.sub_graph_class_defs.append(SubGraphClassDef(name=class_def_name, color=sub_graph_color))
 
     def reset(self):
         self.is_active = False
         self.nx_graph = nx.DiGraph()
         self.start_node = None
-        self.sub_graph_class_defs = []
 
     @property
     def _is_debug_mode(self) -> bool:
@@ -376,169 +324,11 @@ class MissionTracker(MissionTrackerProtocol):
             attributes=edge_attributes,
         )
 
-    def generate_mermaid_flowchart(self, title: Optional[str] = None, subtitle: Optional[str] = None) -> Tuple[str, str]:
-        if not self.is_active:
-            raise JobHistoryError("Job history is not active")
-        nb_nodes = len(self.nx_graph.nodes)
-        if nb_nodes == 0:
-            raise JobHistoryError("Graph has no nodes")
-        log.debug(f"Generating mermaid flowchart for the whole graph which holds {nb_nodes} nodes")
-        mermaid_settings: Dict[str, Any] = {}
-        if title:
-            mermaid_settings["title"] = title
-        history_graph_config = get_config().pipelex.history_graph_config
-        mermaid_settings["config"] = {}
-        if history_graph_config.applied_theme:
-            mermaid_settings["config"]["theme"] = history_graph_config.applied_theme
-        if history_graph_config.applied_layout:
-            mermaid_settings["config"]["layout"] = history_graph_config.applied_layout
-        if history_graph_config.applied_wrapping_width:
-            mermaid_settings["config"]["flowchart"] = {"wrappingWidth": history_graph_config.applied_wrapping_width}
-        mermaid_code = "---\n"
-        mermaid_code += yaml.dump(mermaid_settings)
-        mermaid_code += "---\n"
-        mermaid_code += "flowchart LR\n"
-
-        graph_tree: GraphTree = GraphTree(nodes_by_subgraph={})
-
-        # First pass: Collect nodes into their respective subgraphs
-        for node in self.nx_graph.nodes:
-            node_attributes = self.nx_graph.nodes[node]
-            if not node_attributes:
-                raise JobHistoryError(f"Node attributes are empty for node '{node}'")
-            node_pipe_layer = node_attributes.get(NodeAttributeKey.SUBGRAPH)
-            if not node_pipe_layer:
-                node_pipe_layer = "Unknown"
-            elif not isinstance(node_pipe_layer, str):
-                raise JobHistoryError(f"Node '{node}' has no pipe stack: {node_attributes}")
-
-            sub_graph = node_pipe_layer or "root"
-            # Split sub_graph by "-" and keep the last token
-            if "-" in sub_graph:
-                sub_graph = sub_graph.split("-")[-1]
-            if sub_graph not in graph_tree.nodes_by_subgraph:
-                graph_tree.nodes_by_subgraph[sub_graph] = []
-            graph_tree.nodes_by_subgraph[sub_graph].append(node)
-
-        # pretty_print(graph_tree, title="Graph tree")
-
-        subgraph_lines = self.generate_subgraph_lines(graph_tree)
-        mermaid_code += "\n".join(subgraph_lines)
-        mermaid_code += "\n"
-
-        # Generate subtitle
-        if subtitle:
-            # this is a hack to add something that looks like a subtitle, but it's actually a node with no stroke and no visible link
-            if self.start_node is None:
-                raise JobHistoryError("Start node is not set")
-            mermaid_code += f"""
-    classDef subtitleNodeClass fill:transparent,stroke:#333,stroke-width:0px;
-    __subtitle__["{subtitle}"]
-    class __subtitle__ subtitleNodeClass;
-    __subtitle__ --> {self.start_node}
-    linkStyle 0 stroke:transparent,stroke-width:0px
-"""
-
-        for sub_graph_class_def in self.sub_graph_class_defs:
-            mermaid_code += f"""
-    classDef {sub_graph_class_def.name} fill:{sub_graph_class_def.color},color:#333,stroke:#333;
-"""
-
-        # Generate edges
-        for edge in self.nx_graph.edges(data=True):
-            source, target, edge_data = edge
-            edge_tag: str
-            edge_type = EdgeCategory(edge_data[EdgeAttributeKey.EDGE_CATEGORY])
-            match edge_type:
-                case EdgeCategory.PIPE:
-                    if pipe_code := edge_data.get(EdgeAttributeKey.PIPE_CODE):
-                        edge_tag = nice_edge_tag(pipe_code)
-                        mermaid_code += f"    {source} -- {edge_tag} {history_graph_config.pipe_edge_style} {target}\n"
-                    else:
-                        raise JobHistoryError(f"Pipe edge missing pipe code: {edge_data}")
-                case EdgeCategory.BATCH:
-                    mermaid_code += f"    {source} {history_graph_config.branch_edge_style} {target}\n"
-                case EdgeCategory.AGGREGATE:
-                    mermaid_code += f"    {source} {history_graph_config.aggregate_edge_style} {target}\n"
-                case EdgeCategory.CONDITION:
-                    condition_expression = edge_data.get(EdgeAttributeKey.CONDITION_EXPRESSION)
-                    if not condition_expression:
-                        raise JobHistoryError(f"Condition edge missing condition expression: {edge_data}")
-                    edge_tag = nice_edge_tag(condition_expression)
-                    mermaid_code += f"    {source} {history_graph_config.condition_edge_style} {target}\n"
-                case EdgeCategory.CHOICE:
-                    chosen_pipe_code = edge_data.get(EdgeAttributeKey.CHOSEN_PIPE)
-                    if not chosen_pipe_code:
-                        raise JobHistoryError(f"No chosen pipe code set for edge {source} --- {target}")
-                    edge_tag = nice_edge_tag(chosen_pipe_code)
-                    mermaid_code += f"    {source} -- {edge_tag} {history_graph_config.choice_edge_style} {target}\n"
-
-        url = make_mermaid_url(mermaid_code)
-        return mermaid_code, url
-
-    def generate_subgraph_lines(self, graph_tree: GraphTree) -> List[str]:
-        subgraph_lines: List[str] = []
-        subgraph_class_lines: List[str] = []
-
-        cycle = 0
-        for subgraph_name, nodes in graph_tree.nodes_by_subgraph.items():
-            node_lines: List[str] = []
-            for node in nodes:
-                # log.debug(f"generate_subgraph_lines for node '{node}'")
-                node_attributes = self.nx_graph.nodes[node]
-                if not node_attributes:
-                    raise JobHistoryError(f"Node attributes are empty for node '{node}'")
-                node_category = NodeCategory(node_attributes[NodeAttributeKey.CATEGORY])
-                node_tag = node_attributes[NodeAttributeKey.TAG]
-                if not node_tag:
-                    raise JobHistoryError(f"Node tag is empty for node '{node}'")
-                node_text = node_tag
-                if self._is_debug_mode:
-                    if node_comment := node_attributes.get(NodeAttributeKey.COMMENT):
-                        node_text += f"\n\n{node_comment}"
-                    else:
-                        node_text += "\n\nNo comment"
-                    if node_debug_info := node_attributes.get(NodeAttributeKey.DEBUG_INFO):
-                        node_text += f"\n\n{node_debug_info}"
-                match node_category:
-                    case NodeCategory.SPECIAL:
-                        node_lines.append(f'{node}(["{node_text}"])')
-                    case NodeCategory.STUFF:
-                        node_lines.append(f'{node}["{node_text}"]')
-                    case NodeCategory.CONDITION:
-                        node_lines.append(f'{node}{{"{node_text}"}}')
-                if get_config().pipelex.history_graph_config.is_include_interactivity:
-                    if node_description := node_attributes.get(NodeAttributeKey.DESCRIPTION):
-                        if not isinstance(node_description, str):
-                            raise JobHistoryError(f"Node description is not a string: {node_description}")
-                        node_description = clean_str_for_mermaid_node_title(node_description)
-                        node_lines.append(f'click {node} stuff_node_callback "{node_description}"')
-
-            if not node_lines:
-                raise JobHistoryError(f"No node lines found for subgraph '{subgraph_name}'")
-
-            if subgraph_name == "root":
-                for mermaid_line in node_lines:
-                    subgraph_lines.append(_indent_line(mermaid_line, 2))
-            else:
-                subgraph_lines.append(_indent_line(f'subgraph "{subgraph_name}"', 1))
-                subgraph_lines.append(_indent_line("direction LR", 1))
-                for mermaid_line in node_lines:
-                    subgraph_lines.append(_indent_line(mermaid_line, 2))
-                subgraph_lines.append(_indent_line("end", 1))
-
-                class_def = self.sub_graph_class_defs[cycle % len(self.sub_graph_class_defs)]
-                subgraph_class_lines.append(f"class {subgraph_name} {class_def.name};")
-            cycle += 1
-
-        subgraph_lines.extend(subgraph_class_lines)
-
-        return subgraph_lines
-
     def print_mermaid_flowchart_code_and_url(self, title: Optional[str] = None, subtitle: Optional[str] = None):
-        if not self.is_active:
-            return
-        mermaid_code, url = self.generate_mermaid_flowchart(title=title, subtitle=subtitle)
+        if self.start_node is None:
+            raise JobHistoryError("Start node is not set")
+        flowchart = MissionFlowChart(nx_graph=self.nx_graph, start_node=self.start_node, is_debug_mode=self._is_debug_mode)
+        mermaid_code, url = flowchart.generate_mermaid_flowchart(title=title, subtitle=subtitle)
         print(mermaid_code)
         title_to_print = "Mermaid flowchart URL"
         if title:
@@ -546,9 +336,10 @@ class MissionTracker(MissionTrackerProtocol):
         print_mermaid_url(url=url, title=title_to_print)
 
     def print_mermaid_flowchart_url(self, title: Optional[str] = None, subtitle: Optional[str] = None):
-        if not self.is_active:
-            return
-        _, url = self.generate_mermaid_flowchart(title=title, subtitle=subtitle)
+        if self.start_node is None:
+            raise JobHistoryError("Start node is not set")
+        flowchart = MissionFlowChart(nx_graph=self.nx_graph, start_node=self.start_node, is_debug_mode=self._is_debug_mode)
+        _, url = flowchart.generate_mermaid_flowchart(title=title, subtitle=subtitle)
         title_to_print = "Mermaid flowchart URL"
         if title:
             title_to_print += f" for {title}"
